+++ frontend/app/api/update/meeting/route.ts
@@ -13,12 +13,13 @@ import { lockResourceClaims, ResourceClaim } from "../../../../util/meetings/res
 import { meetingSchema, editScopeSchema, linkedScheduleSchema, LinkedScheduleInput } from "../../../../util/meetings/meetingValidation";
 import {
   linkedFamilyLoader, LinkedFamilyLoader, LinkedFamilyMeeting, getLinkedFamily, familyMembers,
-  canLinkSchedule, availableModesFor, claimedDaysFor, isZoomBearing, LINKED_SCHEDULE_CAP,
+  canLinkSchedule, availableModesFor, claimedDaysFor, isZoomBearing, deriveLinkedScheduleStart,
+  LINKED_SCHEDULE_CAP,
 } from "../../../../util/meetings/linkedSchedules";
 import { isSharedZoomScheduleCompatible } from "../../../../util/meetings/sharedZoomSchedule";
 import { reconcilePendingResume, tearDownPendingResumeSeries } from "../../../../util/meetings/suspension";
-import { calculateEndDateFromOccurrences, firstOccurrenceOnOrAfter } from "../../../../util/meetings/meetingOccurrences";
-import { convertETToUTC, getETTimeOfDay, isConvertETToUTCValidationError } from "../../../../util/date/timeUtils";
+import { calculateEndDateFromOccurrences } from "../../../../util/meetings/meetingOccurrences";
+import { isConvertETToUTCValidationError } from "../../../../util/date/timeUtils";
 import { EditScope, countOccurrencesBefore, exclusionInstant, trimmedEndDate, isLiveOccurrence, rootSplitMid, toETDateStr } from "../../../../util/meetings/editScope";
 import { prisma } from "../../../../lib/prisma";
 
@@ -755,55 +756,6 @@ async function handleScopedEdit(
   );
 
   return NextResponse.json({ ...updatedParent, newMid: createdMeeting.mid });
-}
-
-// The anchor's ET wall-clock time of day and duration, re-anchored onto the first date its
-// series actually meets on `daysOfWeek`. A linked schedule never gets its start from the client:
-// the family is served by ONE Zoom meeting, which can only hold one series, so every member has
-// to keep the same time of day and duration (isSharedZoomScheduleCompatible) -- and the search
-// runs against the ANCHOR's pattern, so an every-other-week family stays in one week phase
-// instead of the two schedules landing on alternating weeks. Returns null when the requested
-// days produce no occurrence inside the anchor's series at all (e.g. a series that ends first).
-function deriveLinkedScheduleStart(
-  anchor: Meeting & { recurrencePattern: RecurrencePattern | null },
-  daysOfWeek: string[],
-): { startDateTime: Date; endDateTime: Date; patternStartDate: Date } | null {
-  const pattern = anchor.recurrencePattern!;
-  // Never earlier than today: a schedule added to a series that started years ago is a schedule
-  // that starts NOW, not one that retroactively met every week since. Searching from the series
-  // start instead would publish a backdated Google Calendar series (fabricated history on past
-  // calendar views) and, for a count-bounded anchor, could resolve an end date that has already
-  // passed -- a row born dead. The week phase is unaffected: matchesRecurrencePattern measures
-  // the interval from the ANCHOR's own startDate below, whatever date the search begins on.
-  const seriesStartEtDateStr = toETDateStr(pattern.startDate);
-  const todayEtDateStr = toETDateStr(new Date());
-  const searchFromEtDateStr = seriesStartEtDateStr > todayEtDateStr ? seriesStartEtDateStr : todayEtDateStr;
-  const firstETDate = firstOccurrenceOnOrAfter(
-    {
-      type: 'weekly',
-      startDate: pattern.startDate,
-      endDate: pattern.endDate,
-      interval: pattern.interval,
-      daysOfWeek,
-      weekOfMonth: null,
-      dayOfMonth: null,
-      // The anchor's own excluded dates are its alone -- a per-occurrence deletion on the
-      // Monday-Friday schedule says nothing about when the Saturday one starts.
-      excludedDates: [],
-    },
-    searchFromEtDateStr,
-  );
-  if (!firstETDate) return null;
-  const { hour, minute, second } = getETTimeOfDay(anchor.startDateTime);
-  const pad = (value: number) => String(value).padStart(2, '0');
-  const startDateTime = new Date(convertETToUTC(`${firstETDate}T${pad(hour)}:${pad(minute)}:${pad(second)}`));
-  const durationMs = anchor.endDateTime.getTime() - anchor.startDateTime.getTime();
-  // INVARIANT: a stored recurrencePattern.startDate is ET-midnight-anchored (parseMMDDYYYY),
-  // never a real meeting-time instant -- calculateEndDateFromOccurrences reads the weekday
-  // anchor straight off getUTCDay(), so a 7 PM-or-later ET start would roll into the next UTC
-  // day and count the occurrences from the wrong weekday.
-  const patternStartDate = new Date(convertETToUTC(`${firstETDate}T00:00:00`));
-  return { startDateTime, endDateTime: new Date(startDateTime.getTime() + durationMs), patternStartDate };
 }
 
 // Runs after the response is sent — the new linked row's own half of the create: its Google
